move all code in a single `lib.rs` file
