--- native/xqlitenif/src/atoms.rs
@@ -1,50 +1,0 @@
-rustler::atoms! {
-    already_closed,
-    atom,
-    binary,
-    cannot_allocate_binary,
-    cannot_close,
-    cannot_close_connection,
-    cannot_convert_atom_to_string,
-    cannot_convert_to_sqlite_value,
-    cannot_decode,
-    cannot_decode_pool_option,
-    cannot_encode,
-    cannot_execute,
-    cannot_execute_pragma,
-    cannot_fetch_row,
-    cannot_open_database,
-    cannot_prepare_statement,
-    cannot_read_column,
-    connection_not_found,
-    db_name,
-    error,
-    expected_keyword_list,
-    expected_keyword_tuple,
-    expected_list,
-    r#false,
-    float,
-    function,
-    integer,
-    invalid_idle_connection_count,
-    invalid_pool_size,
-    invalid_time_value,
-    list,
-    lock_error,
-    map,
-    nil,
-    no_value,
-    ok,
-    pid,
-    port,
-    pragma_get_failed,
-    pragma_put_failed,
-    reference,
-    timeout,
-    r#true,
-    tuple,
-    unknown,
-    unsupported_atom,
-    unsupported_data_type,
-    unsupported_value
-}
